+++ vibeio/src/op/accept.rs
@@ -24,7 +24,7 @@ use windows_sys::Win32::{
 use crate::driver::CompletionIoResult;
 use crate::fd_inner::{InnerRawHandle, RawOsHandle};
 use crate::op::Op;
-use crate::{current_driver, driver::AnyDriver};
+use crate::{driver::AnyDriver, try_current_driver};
 
 #[cfg(unix)]
 fn set_cloexec(fd: RawFd) -> Result<(), io::Error> {
@@ -669,7 +669,7 @@ impl Drop for AcceptOp<'_> {
     #[inline]
     fn drop(&mut self) {
         if let Some(completion_token) = self.completion_token {
-            if let Some(driver) = current_driver() {
+            if let Some(driver) = try_current_driver() {
                 driver.ignore_completion(completion_token, Box::new(()));
             }
         }