+++ vibeio/src/op/fsync.rs
@@ -82,7 +82,7 @@ impl Drop for FsyncOp<'_> {
     #[inline]
     fn drop(&mut self) {
         if let Some(completion_token) = self.completion_token {
-            if let Some(driver) = crate::current_driver() {
+            if let Some(driver) = crate::try_current_driver() {
                 driver.ignore_completion(completion_token, Box::new(()));
             }
         }