+++ vibeio/src/op/read.rs
@@ -308,7 +308,7 @@ impl<B: IoBufMut> Drop for ReadOp<'_, B> {
     #[inline]
     fn drop(&mut self) {
         if let Some(completion_token) = self.completion_token {
-            if let Some(driver) = crate::current_driver() {
+            if let Some(driver) = crate::try_current_driver() {
                 #[cfg(windows)]
                 let completion_state = self.socket_buf.take();
                 #[cfg(not(windows))]