+++ vibeio/src/op/readat.rs
@@ -174,7 +174,7 @@ impl<B: IoBufMut> Drop for ReadAtOp<'_, B> {
     #[inline]
     fn drop(&mut self) {
         if let Some(completion_token) = self.completion_token {
-            if let Some(driver) = crate::current_driver() {
+            if let Some(driver) = crate::try_current_driver() {
                 if let Some(buf) = self.buf.take() {
                     driver.ignore_completion(completion_token, Box::new(buf.into_stable_box()));
                 } else {