+++ vibeio/src/op/recvfrom.rs
@@ -462,7 +462,7 @@ impl<B: IoBufMut> Drop for RecvfromOp<'_, B> {
     #[inline]
     fn drop(&mut self) {
         if let Some(completion_token) = self.completion_token {
-            if let Some(driver) = crate::current_driver() {
+            if let Some(driver) = crate::try_current_driver() {
                 #[cfg(any(windows, target_os = "linux"))]
                 let completion_state = self.completion_state.take();
                 #[cfg(not(any(windows, target_os = "linux")))]