+++ vibeio/src/op/send.rs
@@ -252,7 +252,7 @@ impl<B: IoBuf> Drop for SendOp<'_, B> {
     #[inline]
     fn drop(&mut self) {
         if let Some(completion_token) = self.completion_token {
-            if let Some(driver) = crate::current_driver() {
+            if let Some(driver) = crate::try_current_driver() {
                 #[cfg(windows)]
                 let completion_state = self.socket_buf.take();
                 #[cfg(not(windows))]